refactor(display): change from vuex to pinia
